+++ src/grid/internal/index.ts
@@ -1,0 +1,2 @@
+export * from './NonVirtualGrid';
+export * from './VirtualGrid';