+++ src/grid/GridHead.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 
 export interface HeadCellRenderArgs {
     columnIndex: number;
-    isScrolling?: boolean;
 }
 
 export type HeadCellRender = (args: HeadCellRenderArgs) => React.ReactNode;