+++ src/table/TableRow.tsx
@@ -46,4 +46,4 @@ export class TableRow extends React.PureComponent<TableRowProps> {
 
 }
 
-export type RowRender = (index: number, isScrolling: boolean) => React.SubComp<TableRow> | RowContent;
+export type RowRender = (index: number) => React.SubComp<TableRow> | RowContent;