+++ src/core/VirtualList.tsx
@@ -1,4 +1,4 @@
-// import { VirtualList } from './react-window';
-import { VirtualList } from './recyclerlistview';
+import { VirtualList } from './react-window';
+// import { VirtualList } from './recyclerlistview';
 
 export { VirtualList };